+++ src/application_settings/_private/toml_file_operations.py
@@ -4,17 +4,27 @@ from pathlib import Path
 from typing import Any
 
 import tomlkit
+from loguru import logger
+
+from application_settings._private.file_operations_utils import deep_update
 
 
 def load_toml(path: Path) -> dict[str, Any]:
     """Load the info in the toml file given by path and return as dict"""
     data_stored: dict[str, Any] = {}
-    with path.open(mode="r") as fptr:
-        data_stored = tomlkit.load(fptr)
+    if (
+        path.stat().st_size > 0
+    ):  # this evaluates to false if the file does not exist or is empty
+        with path.open(mode="r") as fptr:
+            data_stored = tomlkit.load(fptr)
+    else:
+        logger.warning(f"File {path} does not exist or is empty.")
     return data_stored
 
 
 def save_toml(path: Path, data: dict[str, Any]) -> None:
-    """Save the info in the data dict to the toml file given by path"""
+    """Update the toml file given by path with data"""
+    old_data = load_toml(path)
+    updated_data = deep_update(old_data, data)
     with path.open(mode="w") as fptr:
-        tomlkit.dump(data, fptr)
+        tomlkit.dump(updated_data, fptr)