+++ src/application_settings/configuring_base.py
@@ -4,10 +4,8 @@ import sys
 from typing import TypeVar
 
 from application_settings.container_base import ContainerBase
-from application_settings.container_section_base import (
-    ContainerSectionBase,
-    SectionTypeStr,
-)
+from application_settings.container_section_base import ContainerSectionBase
+from application_settings.parameter_kind import ParameterKind
 
 from ._private.file_operations import FileFormat
 
@@ -24,21 +22,19 @@ ConfigT.__doc__ = "Represents ConfigBase and all subclasses"
 class ConfigSectionBase(ContainerSectionBase):
     """Base class for all ConfigSection classes (so that we can bound a TypeVar)"""
 
-    @override
     @classmethod
-    def kind_string(cls) -> SectionTypeStr:
-        """Return 'Config'"""
-        return "Config"
+    def kind(cls) -> ParameterKind:
+        """Return ParameterKind.CONFIG"""
+        return ParameterKind.CONFIG
 
 
 class ConfigBase(ContainerBase):
     """Base class for main Config class"""
 
-    @override
     @classmethod
-    def kind_string(cls) -> SectionTypeStr:
-        """Return 'Config'"""
-        return "Config"
+    def kind(cls) -> ParameterKind:
+        """Return ParameterKind.CONFIG"""
+        return ParameterKind.CONFIG
 
     @override
     @classmethod