+++ CHANGELOG.md
@@ -14,10 +14,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Document parameters by means of docstrings according to (the rejected) 
   [PEP224](https://peps.python.org/pep-0224/)
   (Issue [#140](https://github.com/StockwatchDev/application_settings/issues/140))
+- Allow to specify the container class that takes care of loading in the parameter file rather than in
+  code (Issue [#237](https://github.com/StockwatchDev/application_settings/issues/237))
+- Introduced a ParameterKind enum.
 
 ### Changed - 0.5.0
 
 - Use [`tomlkit`](https://tomlkit.readthedocs.io/en/latest/) instead of `tomli` and `tomli-w` (non-breaking)
+- Rename literal SectionTypeStr to ParameterKindStr (breaking).
+
 
 ## [0.4.0] - 2023-12-11